pyproject.toml
[project]
name = "ai-langgraph-reflexion-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "boto3>=1.39.14",
    "fastmcp>=2.10.6",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-aws>=0.2.29",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-ollama>=0.3.6",
    "langgraph>=0.5.4",
    "mcp[cli]>=1.12.2",
    "python-dotenv>=1.1.1",
    "serpapi>=0.1.5",
    "tavily-python>=0.7.10",
]
